Fix decoding tokens having a single scope
Given a configuration having a single scope

  Rails.application.config.middleware.use OmniAuth::Builder do
    provider :eve_online_sso,
      ENV["EVE_ONLINE_SSO_CLIENT_ID"],
      ENV["EVE_ONLINE_SSO_SECRET_KEY"],
      scope: "esi-calendar.read_calendar_events.v1"
  end

Parsing the OAuth token fails with an error

> NoMethodError: undefined method `join' for "esi-calendar.read_calendar_events.v1":String

This is because ESI returns an Array for multiple scopes but a String
for a single one. Yet the code responsible for parsing the token
always expects an Array.

The problem is solved with adapting the code to accommodate for both multiple
and single scopes.

Closes #20